Add TextTable to render beer lists in an attractive manner
The CLI output of cmd/beerlist/beerlist.go was serviceable, but really
ugly. This change adds a TextTable type that will format a tap list in a
nicer-looking format for fixed-width fonts.

The TextTable went into the default beerweb package, and I don't really
think it belongs there, so much, but I'm going to leave it there for
now.

The header() method is extremely ugly and may mutate the state of the
TextTable variable, which sets up gnarly order-dependent calls in
String(), where header() must be called first before we know how wide
the table will actually be. Specifically, at first the underline row was
being generated before header() was called, and if header() changed the
necessary column widths (because a header title was longer than any of
the content in that column), then the underline would be too short.

That needs to be fixed, and there's a much better way to handle this,
but for now, it does what I had intended. I'll fix it next time I touch
it.